Keep snprintf.c as stand-alone module
Source: https://github.com/weiss/c99-snprintf
