+++ Modelica/Resources/BuildProjects/autotools/Makefile.am
@@ -2,7 +2,7 @@ lib_LTLIBRARIES = libzlib.la libModelicaExternalC.la libModelicaMatIO.la libMode
 libModelicaExternalC_la_SOURCES      = ../../C-Sources/ModelicaFFT.c ../../C-Sources/ModelicaInternal.c ../../C-Sources/ModelicaRandom.c ../../C-Sources/ModelicaStrings.c
 libModelicaIO_la_SOURCES             = ../../C-Sources/ModelicaIO.c
 libModelicaIO_la_LIBADD              = libModelicaMatIO.la
-libModelicaMatIO_la_SOURCES          = ../../C-Sources/ModelicaMatIO.c
+libModelicaMatIO_la_SOURCES          = ../../C-Sources/ModelicaMatIO.c ../../C-Sources/snprintf.c
 libModelicaMatIO_la_LIBADD           = libzlib.la @LIBZLIB@ @LIBHDF5@
 libModelicaStandardTables_la_SOURCES = ../../C-Sources/ModelicaStandardTables.c  ../../C-Sources/ModelicaStandardTablesUsertab.c
 libModelicaStandardTables_la_LIBADD  = libModelicaMatIO.la
