+++ Modelica/Resources/BuildProjects/gcc/Makefile
@@ -12,7 +12,8 @@ TABLES_OBJS = \
 	ModelicaStandardTablesUsertab.o
 
 MATIO_OBJS = \
-	ModelicaMatIO.o
+	ModelicaMatIO.o \
+	snprintf.o
 
 IO_OBJS = \
 	ModelicaIO.o
@@ -62,6 +63,9 @@ ModelicaStandardTablesUsertab.o: ../../C-Sources/ModelicaStandardTablesUsertab.c
 
 ModelicaMatIO.o: ../../C-Sources/ModelicaMatIO.c
 	$(CC) $(CPPFLAGS) -DHAVE_ZLIB=1 $(CFLAGS) $(INC) -c -o $@ $<
+
+snprintf.o: ../../C-Sources/snprintf.c
+	$(CC) $(CPPFLAGS) $(CFLAGS) $(INC) -c -o $@ $<
 
 ModelicaIO.o: ../../C-Sources/ModelicaIO.c
 	$(CC) $(CPPFLAGS) $(CFLAGS) $(INC) -c -o $@ $<
